Added the rest of my GUI
In this commit, I have added the rest of my GUI in Products and Coins, which is used to display the GUI in the layout section of the main program. Adding the rest of the products and coins seems to have fixed some of the issues I was having also making the return function work (I am not entirely sure why yet). Overall everything appears to be working as it should be (I think) and I guess I will try to do some of the bonus questions