Updating RPI addition
In this commit, I have added import platform, which will allows me to use platform.system() to see if the computer is running on Widows or Linux. I have chosen to do it this way for now because the PI is running on Linux and the PC is running on Windows. I have also added time so I can open the door for the servo and close it after a certain time.